+++ app/routes/components/CopyUiProvider/changelog.codegen.ts
@@ -5,4 +5,6 @@ export const changelog = ''
   + '- 2025-10-24: Initial version.\n'
   + '- 2025-10-25: Use `useRef` to avoid re-creating zustand store on each render.\n'
   + '- 2025-10-27: Wrap children with `Toast.Context`.\n'
+  + '- 2025-11-04: Move the original `ThemeProvider` into `CopyUiProvider` and\n'
+  + '  support the `useTheme`, etc. The `ThemeProvider` now is deprecated.\n'
   ;