+++ src/components/CopyUiProvider/CHANGELOG.md
@@ -1,3 +1,5 @@
 - 2025-10-24: Initial version.
 - 2025-10-25: Use `useRef` to avoid re-creating zustand store on each render.
 - 2025-10-27: Wrap children with `Toast.Context`.
+- 2025-11-04: Move the original `ThemeProvider` into `CopyUiProvider` and
+  support the `useTheme`, etc. The `ThemeProvider` now is deprecated.